requirements.txt
psutil
pywin32
pynput
pystray
Pillow
plyer
matplotlib